--not work--

+++ dotNet2022_8090_7731/DalXml/DalXml.cs
@@ -199,11 +199,11 @@ namespace Dal
         public (double, double, double, double, double) PowerConsumptionRequest()
         {
             XDocument document = XDocument.Load(configFilePath);
-            double Available = int.Parse(document.Root.Element("Available").Value);
-            double LightWeight = int.Parse(document.Root.Element("LightWeight").Value);
-            double MediumWeight = int.Parse(document.Root.Element("MediumWeight").Value);
-            double HeavyWeight = int.Parse(document.Root.Element("HeavyWeight").Value);
-            double ChargingRate = int.Parse(document.Root.Element("ChargingRate").Value);
+            double Available = double.Parse(document.Root.Element("Available").Value);
+            double LightWeight = double.Parse(document.Root.Element("LightWeight").Value);
+            double MediumWeight = double.Parse(document.Root.Element("MediumWeight").Value);
+            double HeavyWeight = double.Parse(document.Root.Element("HeavyWeight").Value);
+            double ChargingRate = double.Parse(document.Root.Element("ChargingRate").Value);
 
             document.Save(configFilePath);
             return (Available,
